requirements.txt
pygame
pillow
cx_freeze